+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- XMLParser (pheme/parser/xml.py)
 ### Changed
 ### Deprecated
 ### Removed